Bunch of changes based on PR code review.
The unit tests are failing now but it is in the code that prepares the workflow, so I think the problem is with the code I merged in from the master branch which I haven't checked how is changed. @dansondergaard, can you see why there is a key error in a dictionary lookup in the workflow preparation now?