Refactor backends to make them more testable.
Along with this commit follows a multitude of fixes for both the slurm
and local backend.

Also, backends now have a single Backend.status() method
which returns the status of a target. This means that users only need
to call the backend once to know the status of a target. Previously the
user had to call running(), then submitted(), then completed(), which
was stupid, especially if we end up calling e.g. the slurm queue manager
when the status is requested for a target.