Fix bug in scheduling.
The bug was revealed since I recently made backends more strict when
submitting targets. Previously, when submitting a target to a backend
with a dependency that wasn't tracked by the backend (which should never
happen), the untracked dependency would just be ignore. Now, backends
require the dependencies to be tracked (i.e. have been submitted
before).

However, the scheduling code did not take this into account and simply
handed over a full list of dependencies (not only the dependencies
that were submitted earlier). This caused a lookup error in the backend.

The scheduling code has been rewritten to be recursive. It keeps track
of the total schedule (all targets that have been submitted), but also
keeps track of which direct dependencies of a target that have been
submitted. This list is passed to the backend.