Fix bug when loading workflow and running Python 3.4.
This release removes compatibility for Python 3.4 since the `imp` module
causes a significant amount of weird behaviour. There's no reason to use
3.4 over 3.5, so this was an easy decision.

The bug was demonstrated by the example workflow included in this
commit and caused gwf to claim that the same target was defined twice,
since the 'gwf' object was mysteriously being redefined when importing
the other workflow. Using the importlib module provides substantially
saner behaviour when importing modules dynamically (e.g. it doesn't mess
with sys.path) and thus solves the bug.