Move file checking from Graph to Scheduler.
Previously, the Graph would check whether a target depended on a file
which wasn't provided by any other target. If this was the case, the
Graph would check whether the file existed on disk, and if not, raise
an excaption.

With this change the Graph doesn't check whether the file exists on
disk, but just registers the file as "unresolved".

This means that a Graph doesn't do any IO at all, which makes testing
easier. It also means that IO is only performed by the Scheduler, right
before scheduling a target, which makes some actions faster.