Refactoring of backends
So far the SGE backend was basically a copy of the Slurm backend with a
few modifications. This commit introduces a PbsLikeBackendBase which
contains all of the shared functionality of PBS-like backends.
SlurmBackend and SgeBackend now inherits from this class.

The Backend class was previously backed by a metaclass which checked for
required methods and wrapped the submit() method in a function which
ensured that target options were properly set. The metaclass and wrapper
has now been removed and targets are now prepared by the scheduler
before they're submitted to the backend.