Scheduler improvements (#324)
Make the scheduler side-effect free

The scheduler is now (in some sense) a pure function which returns which
targets were scheduled and thus should be submitted to the backend. This
makes testing of the scheduler a lot easier and makes it easier to use.

This is a major change that required rewriting of some tests, as well as
code related to filtering, the status command and the run command.